Added internationalization.
This will use a spreadsheet (currently mine, but for the prod version we should use one owned by the project owner) to look up links to the different sheets.
So adding a country will be as "easy" as adding the correct links to the sheet (link: https://docs.google.com/spreadsheets/d/1DmnjXEUCGI49o4a60zhrFf2Uq6EJ2x2WjSletnadrFs/edit#gid=0 ) and done.

I've tried to test this as good as I can, I hope it will at least inspire someone to make this internationalization happen (^_^)

I'm running this version on my own PC (on Chrome with Tampermonkey) and it works for the Netherlands now also.
There are still things that need to be changed to have all the features adapted for different countries (like phone number check and the opening hours parse)
because I've hard coded those into this version to support Dutch terms as well.